Task 3 finished

+++ src/main/java/com/mindhub/homebanking/controllers/AccountController.java
@@ -23,7 +23,7 @@ public class AccountController {
     }
 
     @RequestMapping("/accounts/{id}")
-    public AccountDTO getAccount(@PathVariable Integer id){
+    public AccountDTO getAccount(@PathVariable Long id){
         return accountRepository.findById(id).map(AccountDTO::new).orElse(null);
     }
 }
